[Android][Fixed] Break the circular dependency
