+++ ReactAndroid/src/main/java/com/facebook/react/CoreModulesPackage.java
@@ -29,6 +29,7 @@ import com.facebook.react.modules.systeminfo.AndroidInfoModule;
 import com.facebook.react.uimanager.UIImplementationProvider;
 import com.facebook.react.uimanager.UIManagerModule;
 import com.facebook.react.uimanager.ViewManager;
+import com.facebook.react.uimanager.ViewManagerResolver;
 import com.facebook.systrace.Systrace;
 import java.util.Arrays;
 import java.util.List;
@@ -149,7 +150,7 @@ import javax.inject.Provider;
     Systrace.beginSection(Systrace.TRACE_TAG_REACT_JAVA_BRIDGE, "createUIManagerModule");
     try {
       if (mLazyViewManagersEnabled) {
-        UIManagerModule.ViewManagerResolver resolver = new UIManagerModule.ViewManagerResolver() {
+        ViewManagerResolver resolver = new ViewManagerResolver() {
           @Override
           public @Nullable ViewManager getViewManager(String viewManagerName) {
             return mReactInstanceManager.createViewManager(viewManagerName);
