+++ ReactAndroid/src/main/java/com/facebook/react/uimanager/BUCK
@@ -38,13 +38,13 @@ rn_android_library(
         react_native_target("java/com/facebook/react/uimanager/annotations:annotations"),
         react_native_target("java/com/facebook/react/uimanager/util:util"),
         react_native_target("java/com/facebook/react/uimanager/common:common"),
-        react_native_target("java/com/facebook/react/views/text:text_for_uimanager"),
         react_native_target("res:uimanager"),
     ],
     exported_deps = [
         ":DisplayMetrics",
         react_native_dep("third-party/java/jsr-305:jsr-305"),
         react_native_target("java/com/facebook/react/uimanager/common:common"),
+        react_native_target("java/com/facebook/react/views/text:text_for_uimanager"),
     ],
 )
 
@@ -66,25 +66,135 @@ rn_android_library(
 
 rn_android_library(
     name = "uimanager_for_text",
-    srcs = glob([
-        "ReactShadowNode.java",
-        "ReactShadowNodeImpl.java",
-        "annotations/ReactProp.java",
-        "IllegalViewOperationException.java",
-        "LayoutShadowNode.java",
-        "PixelUtil.java",
-        "ReactShadowNode.java",
-        "ViewDefaults.java",
-        "ViewProps.java",
-    ]),
+    srcs = glob(
+        [
+            "debug/*.java",
+            "events/*.java",
+            "layoutanimation/*.java",
+            "PixelUtil.java",
+            "RootViewUtil.java",
+            "IllegalViewOperationException.java",
+            "RootView.java",
+            "LayoutShadowNode.java",
+            "ReactShadowNode.java",
+            "ViewDefaults.java",
+            "ViewProps.java",
+            "ReactShadowNodeImpl.java",
+            "ReactStylesDiffMap.java",
+            "Spacing.java",
+            "UIViewOperationQueue.java",
+            "NativeViewHierarchyOptimizer.java",
+            "ThemedReactContext.java",
+            "NativeViewHierarchyManager.java",
+            "GuardedFrameCallback.java",
+            "ViewAtIndex.java",
+            "LayoutUpdateListener.java",
+            "UIBlock.java",
+            "ShadowNodeRegistry.java",
+            "ViewManager.java",
+            "ViewManagerRegistry.java",
+            "RootViewManager.java",
+            "ViewGroupManager.java",
+            "ViewManagerResolver.java",
+            "BaseViewManager.java",
+            "AccessibilityDelegateUtil.java",
+            "MatrixMathHelper.java",
+            "ViewManagerPropertyUpdater.java",
+            "ViewManagersPropertyCache.java",
+            "ReactZIndexedViewGroup.java",
+            "AccessibilityHelper.java",
+            "TransformHelper.java",
+            "ReactYogaConfigProvider.java",
+            "YogaNodePool.java",
+            "NoSuchNativeViewException.java",
+            "FloatUtil.java",
+            "TouchTargetHelper.java",
+            "UIManagerModuleConstants.java",
+            "PointerEvents.java",
+            "ReactPointerEventsView.java",
+            "ReactCompoundView.java",
+            "ReactCompoundViewGroup.java",
+        ],
+        exclude = ["DisplayMetricsHolder.java"],
+    ),
+    provided_deps = [
+        react_native_dep("third-party/android/support/v4:lib-support-v4"),
+        react_native_dep("third-party/android/support-annotations:android-support-annotations"),
+    ],
     required_for_source_only_abi = True,
     visibility = [
         "PUBLIC",
     ],
     deps = [
         YOGA_TARGET,
+        ":DisplayMetrics",
+        react_native_dep("java/com/facebook/systrace:systrace"),
+        react_native_dep("libraries/fbcore/src/main/java/com/facebook/common/logging:logging"),
         react_native_dep("third-party/java/infer-annotations:infer-annotations"),
-        react_native_target("java/com/facebook/react/uimanager/annotations:annotations"),
+        react_native_target("java/com/facebook/debug/tags:tags"),
+        react_native_target("java/com/facebook/debug/holder:holder"),
+        react_native_target("java/com/facebook/react/animation:animation"),
+        react_native_target("java/com/facebook/react/bridge:bridge"),
+        react_native_target("java/com/facebook/react/common:common"),
+        react_native_target("java/com/facebook/react/config:config"),
+        react_native_target("java/com/facebook/react/module/annotations:annotations"),
+        react_native_target("java/com/facebook/react/modules/core:core"),
         react_native_target("java/com/facebook/react/modules/i18nmanager:i18nmanager"),
+        react_native_target("java/com/facebook/react/touch:touch"),
+        react_native_target("java/com/facebook/react/uimanager/annotations:annotations"),
+        react_native_target("java/com/facebook/react/uimanager/util:util"),
+        react_native_target("java/com/facebook/react/uimanager/common:common"),
+        react_native_target("res:uimanager"),
+    ],
+    exported_deps = [
+        ":DisplayMetrics",
+        react_native_dep("third-party/java/jsr-305:jsr-305"),
+        react_native_target("java/com/facebook/react/uimanager/common:common"),
     ],
 )
+
+rn_android_library(
+    name = "uimanager_for_view",
+    srcs = glob(
+        [
+            "FloatUtil.java",
+            "Spacing.java",
+        ],
+        exclude = ["DisplayMetricsHolder.java"],
+    ),
+    provided_deps = [
+        react_native_dep("third-party/android/support/v4:lib-support-v4"),
+        react_native_dep("third-party/android/support-annotations:android-support-annotations"),
+    ],
+    required_for_source_only_abi = True,
+    visibility = [
+        "PUBLIC",
+    ],
+    deps = [
+        YOGA_TARGET,
+        ":DisplayMetrics",
+        react_native_dep("java/com/facebook/systrace:systrace"),
+        react_native_dep("libraries/fbcore/src/main/java/com/facebook/common/logging:logging"),
+        react_native_dep("third-party/java/infer-annotations:infer-annotations"),
+        react_native_target("java/com/facebook/debug/tags:tags"),
+        react_native_target("java/com/facebook/debug/holder:holder"),
+        react_native_target("java/com/facebook/react/animation:animation"),
+        react_native_target("java/com/facebook/react/bridge:bridge"),
+        react_native_target("java/com/facebook/react/common:common"),
+        react_native_target("java/com/facebook/react/config:config"),
+        react_native_target("java/com/facebook/react/module/annotations:annotations"),
+        react_native_target("java/com/facebook/react/modules/core:core"),
+        react_native_target("java/com/facebook/react/modules/i18nmanager:i18nmanager"),
+        react_native_target("java/com/facebook/react/touch:touch"),
+        react_native_target("java/com/facebook/react/uimanager/annotations:annotations"),
+        react_native_target("java/com/facebook/react/uimanager/util:util"),
+        react_native_target("java/com/facebook/react/uimanager/common:common"),
+        react_native_target("res:uimanager"),
+    ],
+    exported_deps = [
+        ":DisplayMetrics",
+        react_native_dep("third-party/java/jsr-305:jsr-305"),
+        react_native_target("java/com/facebook/react/uimanager/common:common"),
+    ],
+)
+
