+++ ReactAndroid/src/main/java/com/facebook/react/uimanager/UIViewOperationQueue.java
@@ -96,38 +96,38 @@ public class UIViewOperationQueue {
     }
   }
 
-  private final class EmitOnLayoutEventOperation extends ViewOperation {
-
-    private final int mScreenX;
-    private final int mScreenY;
-    private final int mScreenWidth;
-    private final int mScreenHeight;
-
-    public EmitOnLayoutEventOperation(
-        int tag,
-        int screenX,
-        int screenY,
-        int screenWidth,
-        int screenHeight) {
-      super(tag);
-      mScreenX = screenX;
-      mScreenY = screenY;
-      mScreenWidth = screenWidth;
-      mScreenHeight = screenHeight;
-    }
-
-    @Override
-    public void execute() {
-      mReactApplicationContext.getNativeModule(UIManagerModule.class)
-        .getEventDispatcher()
-        .dispatchEvent(OnLayoutEvent.obtain(
-          mTag,
-          mScreenX,
-          mScreenY,
-          mScreenWidth,
-          mScreenHeight));
-    }
-  }
+//  private final class EmitOnLayoutEventOperation extends ViewOperation {
+//
+//    private final int mScreenX;
+//    private final int mScreenY;
+//    private final int mScreenWidth;
+//    private final int mScreenHeight;
+//
+//    public EmitOnLayoutEventOperation(
+//        int tag,
+//        int screenX,
+//        int screenY,
+//        int screenWidth,
+//        int screenHeight) {
+//      super(tag);
+//      mScreenX = screenX;
+//      mScreenY = screenY;
+//      mScreenWidth = screenWidth;
+//      mScreenHeight = screenHeight;
+//    }
+//
+//    @Override
+//    public void execute() {
+//      mReactApplicationContext.getNativeModule(UIManagerModule.class)
+//        .getEventDispatcher()
+//        .dispatchEvent(OnLayoutEvent.obtain(
+//          mTag,
+//          mScreenX,
+//          mScreenY,
+//          mScreenWidth,
+//          mScreenHeight));
+//    }
+//  }
 
   private final class UpdateInstanceHandleOperation extends ViewOperation {
 
@@ -559,9 +559,9 @@ public class UIViewOperationQueue {
   private final class LayoutUpdateFinishedOperation implements UIOperation {
 
     private final ReactShadowNode mNode;
-    private final UIImplementation.LayoutUpdateListener mListener;
+    private final LayoutUpdateListener mListener;
 
-    private LayoutUpdateFinishedOperation(ReactShadowNode node, UIImplementation.LayoutUpdateListener listener) {
+    private LayoutUpdateFinishedOperation(ReactShadowNode node, LayoutUpdateListener listener) {
       mNode = node;
       mListener = listener;
     }
@@ -756,14 +756,14 @@ public class UIViewOperationQueue {
     mOperations.add(new UpdatePropertiesOperation(reactTag, props));
   }
 
-  public void enqueueOnLayoutEvent(
-    int tag,
-    int screenX,
-    int screenY,
-    int screenWidth,
-    int screenHeight) {
-    mOperations.add(new EmitOnLayoutEventOperation(tag, screenX, screenY, screenWidth, screenHeight));
-  }
+//  public void enqueueOnLayoutEvent(
+//    int tag,
+//    int screenX,
+//    int screenY,
+//    int screenWidth,
+//    int screenHeight) {
+//    mOperations.add(new EmitOnLayoutEventOperation(tag, screenX, screenY, screenWidth, screenHeight));
+//  }
 
 
   public void enqueueUpdateLayout(
@@ -847,7 +847,7 @@ public class UIViewOperationQueue {
     mOperations.add(new SendAccessibilityEvent(tag, eventType));
   }
 
-  public void enqueueLayoutUpdateFinished(ReactShadowNode node, UIImplementation.LayoutUpdateListener listener) {
+  public void enqueueLayoutUpdateFinished(ReactShadowNode node, LayoutUpdateListener listener) {
     mOperations.add(new LayoutUpdateFinishedOperation(node, listener));
   }
